refactor(#741): refactor descriptionController
